Handle exceptions related to invalid event message format.
